nrf_rpc: rpc_utils: reorganize nRF RPC utility commands
nRF RPC device information command library has recently been
generalized into nRF RPC utility command library that
includes commands that do not fit other command groups.

Complete the effort by splitting the implementation so that
commands that are part of the library can be enabled or
disabled independently. Also, improve the CMake and Kconfig
structure so that adding new commands to the library takes
less effort.

Also:
1. Update documentation of nRF RPC utils.
2. Update unit tests so that tests for other commands from
   this library can easily be added.
3. Improve remote shell unit tests by enabling the actual
   shell command.
