--- doc/nrf/libraries/nrf_rpc/nrf_rpc_dev_info.rst
@@ -1,24 +1,0 @@
-.. _nrf_rpc_dev_info:
-
-nRF RPC device information
-##########################
-
-.. contents::
-   :local:
-   :depth: 2
-
-This library allows you to obtain the device information functions of the :ref:`nrfxlib:nrf_rpc` on a remote server.
-
-Configuration
-*************
-
-Use the :kconfig:option:`CONFIG_NRF_RPC_DEV_INFO` Kconfig option to enable this library.
-To enable it for the client, use the :kconfig:option:`CONFIG_NRF_RPC_DEV_INFO_CLIENT` Kconfig option and respectively,
-:kconfig:option:`CONFIG_NRF_RPC_DEV_INFO_SERVER` for the server.
-
-API documentation
-*****************
-
-| Header file: :file:`include/nrf_rpc/nrf_rpc_dev_info.h`
-
-.. doxygengroup:: nrf_rpc_dev_info
