+++ samples/nrf_rpc/protocols_serialization/client/CMakeLists.txt
@@ -14,13 +14,10 @@ target_sources(app PRIVATE
     src/main.c
 )
 
-if(DEFINED CONFIG_NRF_RPC_DEV_INFO OR DEFINED CONFIG_NRF_RPC_CRASH_GEN)
-  zephyr_library_sources(src/rpc_utils_shell.c)
-endif()
-
 zephyr_library_sources_ifdef(CONFIG_OPENTHREAD_RPC src/ot_shell.c)
 zephyr_library_sources_ifdef(CONFIG_MPSL_CX_SOFTWARE src/coex_shell.c)
 zephyr_library_sources_ifdef(CONFIG_LOG_FORWARDER_RPC src/log_rpc_shell.c)
 zephyr_library_sources_ifdef(CONFIG_NFC_RPC src/nfc_shell.c)
 zephyr_library_sources_ifdef(CONFIG_BT_THROUGHPUT src/bt_throughput_shell.c)
+zephyr_library_sources_ifdef(CONFIG_NRF_RPC_UTILS src/rpc_utils_shell.c)
 # NORDIC SDK APP START
