+++ samples/nrf_rpc/protocols_serialization/client/prj.conf
@@ -24,4 +24,7 @@ CONFIG_LOG_FUNC_NAME_PREFIX_DBG=n
 CONFIG_LOG_BACKEND_RTT=n
 CONFIG_SHELL_LOG_BACKEND=y
 
-CONFIG_NRF_RPC_DEV_INFO=y
+CONFIG_NRF_RPC_UTILS=y
+CONFIG_NRF_RPC_UTILS_CLIENT=y
+CONFIG_NRF_RPC_UTILS_DEV_INFO=y
+CONFIG_NRF_RPC_UTILS_REMOTE_SHELL=y
