+++ samples/nrf_rpc/protocols_serialization/client/src/rpc_utils_shell.c
@@ -5,9 +5,11 @@
  */
 
 #include <zephyr/shell/shell.h>
-#include <nrf_rpc/nrf_rpc_dev_info.h>
+#include <nrf_rpc/rpc_utils/crash_gen.h>
+#include <nrf_rpc/rpc_utils/dev_info.h>
+#include <nrf_rpc/rpc_utils/remote_shell.h>
 
-#if defined(CONFIG_NRF_RPC_DEV_INFO)
+#if defined(CONFIG_NRF_RPC_UTILS_DEV_INFO)
 static int remote_version_cmd(const struct shell *sh, size_t argc, char *argv[])
 {
 	char *version = nrf_rpc_get_ncs_commit_sha();
@@ -23,7 +25,9 @@ static int remote_version_cmd(const struct shell *sh, size_t argc, char *argv[])
 
 	return 0;
 }
+#endif
 
+#if defined(CONFIG_NRF_RPC_UTILS_REMOTE_SHELL)
 static int remote_shell_cmd(const struct shell *sh, size_t argc, char *argv[])
 {
 	char *output = nrf_rpc_invoke_remote_shell_cmd(argc - 1, argv + 1);
@@ -39,9 +43,9 @@ static int remote_shell_cmd(const struct shell *sh, size_t argc, char *argv[])
 
 	return 0;
 }
-#endif /* CONFIG_NRF_RPC_DEV_INFO */
+#endif /* CONFIG_NRF_RPC_UTILS_REMOTE_SHELL */
 
-#if defined(CONFIG_NRF_RPC_CRASH_GEN)
+#if defined(CONFIG_NRF_RPC_UTILS_CRASH_GEN)
 static int cmd_assert(const struct shell *sh, size_t argc, char *argv[])
 {
 	int rc = 0;
@@ -98,22 +102,21 @@ static int cmd_stack_overflow(const struct shell *sh, size_t argc, char *argv[])
 
 	return 0;
 }
-#endif /* CONFIG_NRF_RPC_CRASH_GEN */
+#endif /* CONFIG_NRF_RPC_UTILS_CRASH_GEN */
 
-SHELL_STATIC_SUBCMD_SET_CREATE(util_cmds,
-			       #if defined(CONFIG_NRF_RPC_DEV_INFO)
-			       SHELL_CMD_ARG(remote_version, NULL, "Get server version",
-					     remote_version_cmd, 1, 0),
-			       SHELL_CMD_ARG(remote_shell, NULL, "Call remote shell command",
-					     remote_shell_cmd, 2, 255),
-			       #endif
-			       #if defined(CONFIG_NRF_RPC_CRASH_GEN)
-			       SHELL_CMD_ARG(assert, NULL, "Invoke assert", cmd_assert, 1, 1),
-			       SHELL_CMD_ARG(hard_fault, NULL, "Invoke hard fault", cmd_hard_fault,
-					     1, 1),
-			       SHELL_CMD_ARG(stack_overflow, NULL, "Invoke stack overflow",
-					     cmd_stack_overflow, 1, 1),
-			       #endif
-			       SHELL_SUBCMD_SET_END);
+SHELL_STATIC_SUBCMD_SET_CREATE(
+	util_cmds,
+#if defined(CONFIG_NRF_RPC_UTILS_DEV_INFO)
+	SHELL_CMD_ARG(remote_version, NULL, "Get server version", remote_version_cmd, 1, 0),
+#endif
+#if defined(CONFIG_NRF_RPC_UTILS_REMOTE_SHELL)
+	SHELL_CMD_ARG(remote_shell, NULL, "Call remote shell command", remote_shell_cmd, 2, 255),
+#endif
+#if defined(CONFIG_NRF_RPC_UTILS_CRASH_GEN)
+	SHELL_CMD_ARG(assert, NULL, "Invoke assert", cmd_assert, 1, 1),
+	SHELL_CMD_ARG(hard_fault, NULL, "Invoke hard fault", cmd_hard_fault, 1, 1),
+	SHELL_CMD_ARG(stack_overflow, NULL, "Invoke stack overflow", cmd_stack_overflow, 1, 1),
+#endif
+	SHELL_SUBCMD_SET_END);
 
 SHELL_CMD_ARG_REGISTER(rpc, &util_cmds, "nRF RPC utility commands", NULL, 1, 0);
