+++ samples/nrf_rpc/protocols_serialization/server/Kconfig
@@ -37,7 +37,7 @@ config NVS
 	default y if !(SOC_FLASH_NRF_RRAM || SOC_FLASH_NRF_MRAM)
 
 config SHELL_BACKEND_SERIAL
-	default n if NRF_RPC_REMOTE_SHELL
+	default n if NRF_RPC_UTILS_REMOTE_SHELL
 
 if LOG
