+++ samples/nrf_rpc/protocols_serialization/server/prj.conf
@@ -16,8 +16,10 @@ CONFIG_NRF_RPC_UART_RELIABLE=y
 CONFIG_UART_LINE_CTRL=y
 CONFIG_UART_INTERRUPT_DRIVEN=y
 
-CONFIG_NRF_RPC_DEV_INFO=y
-CONFIG_NRF_RPC_DEV_INFO_SERVER=y
+CONFIG_NRF_RPC_UTILS=y
+CONFIG_NRF_RPC_UTILS_SERVER=y
+CONFIG_NRF_RPC_UTILS_DEV_INFO=y
+CONFIG_NRF_RPC_UTILS_REMOTE_SHELL=y
 
 CONFIG_SETTINGS=y
 CONFIG_FLASH=y
