+++ subsys/nrf_rpc/CMakeLists.txt
@@ -18,6 +18,4 @@ zephyr_library_sources_ifdef(CONFIG_NRF_RPC_CALLBACK_PROXY nrf_rpc_cbkproxy.c)
 
 zephyr_library_sources_ifdef(CONFIG_NRF_RPC_UART_TRANSPORT nrf_rpc_uart.c)
 
-if(DEFINED CONFIG_NRF_RPC_DEV_INFO OR DEFINED CONFIG_NRF_RPC_CRASH_GEN)
-  add_subdirectory(rpc_utils)
-endif()
+add_subdirectory_ifdef(CONFIG_NRF_RPC_UTILS rpc_utils)
