+++ tests/subsys/nrf_rpc/rpc_utils/client/CMakeLists.txt
