+++ tests/subsys/nrf_rpc/rpc_utils/client/prj.conf
@@ -16,6 +16,7 @@ CONFIG_MOCK_NRF_RPC_TRANSPORT=y
 
 CONFIG_KERNEL_MEM_POOL=y
 CONFIG_HEAP_MEM_POOL_SIZE=4096
-CONFIG_NRF_RPC_DEV_INFO=y
-CONFIG_NRF_RPC_DEV_INFO_SERVER=y
-CONFIG_NRF_RPC_REMOTE_SHELL=n
+CONFIG_NRF_RPC_UTILS=y
+CONFIG_NRF_RPC_UTILS_CLIENT=y
+CONFIG_NRF_RPC_UTILS_DEV_INFO=y
+CONFIG_NRF_RPC_UTILS_REMOTE_SHELL=y
