+++ tests/subsys/nrf_rpc/rpc_utils/common/nrf_rpc_single_thread.c
