+++ tests/subsys/nrf_rpc/rpc_utils/common/test_rpc_env.h
