+++ tests/subsys/nrf_rpc/rpc_utils/server/CMakeLists.txt
