+++ tests/subsys/nrf_rpc/rpc_utils/server/prj.conf
@@ -16,4 +16,9 @@ CONFIG_MOCK_NRF_RPC_TRANSPORT=y
 
 CONFIG_KERNEL_MEM_POOL=y
 CONFIG_HEAP_MEM_POOL_SIZE=4096
-CONFIG_NRF_RPC_DEV_INFO=y
+CONFIG_NRF_RPC_UTILS=y
+CONFIG_NRF_RPC_UTILS_SERVER=y
+CONFIG_NRF_RPC_UTILS_DEV_INFO=y
+CONFIG_NRF_RPC_UTILS_REMOTE_SHELL=y
+CONFIG_SHELL_BACKEND_SERIAL=n
+CONFIG_SHELL_AUTOSTART=n
